Move data models to sub-package
