+++ src/main/java/com/github/jbence1994/erp/inventory/model/Supplier.java
@@ -1,4 +1,4 @@
-package com.github.jbence1994.erp.inventory;
+package com.github.jbence1994.erp.inventory.model;
 
 public class Supplier {
     private Long id;
